refactor: Grouping to packages.
