+++ adventures.and.beyond/src/object/interactable/objects/TreasureObject.java
@@ -1,4 +1,4 @@
-package object;
+package object.interactable.objects;
 
 import entity.Player;
 import sound.SoundKey;
